Make xdelta1 generation warning-clean
Generate typed serialization registry adapters, make EDSIO regeneration reproducible, and fix the remaining handwritten warnings so Wall, Wextra, and Wpedantic can be enforced.


Copilot-Session: ab638c7d-6c1e-4ffa-9f2e-dbb22624fe48

+++ xdelta1/getopt.c
@@ -474,7 +474,7 @@ _getopt_internal (int argc, char *const *argv, const char *optstring,
       for (p = longopts, option_index = 0; p->name; p++, option_index++)
 	if (!strncmp (p->name, nextchar, nameend - nextchar))
 	  {
-	    if (nameend - nextchar == strlen (p->name))
+	    if ((size_t) (nameend - nextchar) == strlen (p->name))
 	      {
 		/* Exact match found.  */
 		pfound = p;
